requirements.txt
flask
flask-cors
python-dotenv
huggingface_hub
flask-limiter
PyPDF2
gunicorn
